WIP: pristine table triggers support.

+++ svnkit/src/main/java/org/tmatesoft/svn/core/internal/wc17/db/SVNWCDb.java
@@ -1297,6 +1297,7 @@ public class SVNWCDb implements ISVNWCDb {
         final DirParsedInfo parsed = parseDir(wriAbspath, Mode.ReadWrite);
         SVNWCDbDir pdh = parsed.wcDbDir;
         verifyDirUsable(pdh);
+        
         File pristineAbspath = getPristineFileName(pdh, sha1Checksum, true);
         SVNNodeKind kind = SVNFileType.getNodeKind(SVNFileType.getType(pristineAbspath));
         if (kind == SVNNodeKind.FILE) {
@@ -2925,41 +2926,27 @@ public class SVNWCDb implements ISVNWCDb {
         DirParsedInfo parseDir = parseDir(wcRootAbsPath, Mode.ReadWrite);
         SVNWCDbDir pdh = parseDir.wcDbDir;
         verifyDirUsable(pdh);
-        {
-            SVNSqlJetStatement stmt = pdh.getWCRoot().getSDb().getStatement(SVNWCDbStatements.LOOK_FOR_WORK);
-            boolean haveRow;
-            try {
-                haveRow = stmt.next();
-            } finally {
-                stmt.reset();
-            }
-            if (haveRow) {
-                return;
-            }
+        SVNSqlJetStatement stmt = pdh.getWCRoot().getSDb().getStatement(SVNWCDbStatements.LOOK_FOR_WORK);
+        boolean haveRow;
+        try {
+            haveRow = stmt.next();
+        } finally {
+            stmt.reset();
         }
-        boolean isReferenced;
-        {
-            SvnChecksum md5Checksum = getPristineMD5(wcRootAbsPath, sha1Checksum);
-            SVNSqlJetStatement stmt = pdh.getWCRoot().getSDb().getStatement(SVNWCDbStatements.SELECT_ANY_PRISTINE_REFERENCE);
-            try {
-                stmt.bindChecksum(1, sha1Checksum);
-                stmt.bindChecksum(2, md5Checksum);
-                isReferenced = stmt.next();
-            } finally {
-                stmt.reset();
-            }
+        if (haveRow) {
+            return;
         }
-        if (!isReferenced) {
-            pristineRemove(pdh, sha1Checksum);
-        }
+        pristineRemove(pdh, sha1Checksum);
     }
 
     private void pristineRemove(SVNWCDbDir pdh, SvnChecksum sha1Checksum) throws SVNException {
         SVNSqlJetStatement stmt = pdh.getWCRoot().getSDb().getStatement(SVNWCDbStatements.DELETE_PRISTINE);
         stmt.bindChecksum(1, sha1Checksum);
-        stmt.done();
-        File pristineAbspath = getPristineFileName(pdh, sha1Checksum, true);
-        SVNFileUtil.deleteFile(pristineAbspath);
+        
+        if (stmt.done() != 0) {
+            File pristineAbspath = getPristineFileName(pdh, sha1Checksum, true);
+            SVNFileUtil.deleteFile(pristineAbspath);
+        }
     }
 
     public void removeWCLock(File localAbspath) throws SVNException {
